inner_module_def_datastruct/requirements.txt
# dateutil
python-dateutil==2.8.2